Updated: restricted action

+++ Http/Controllers/Backend/UsersController.php
@@ -285,6 +285,14 @@ class UsersController extends Controller
         }
 
         try {
+            $is_restricted = User::restrictedActions('delete', $id);
+
+            if(isset($is_restricted['success']) && !$is_restricted['success'])
+            {
+                $response['success'] = false;
+                $response['errors'] = $is_restricted['errors'];
+                return response()->json($response);
+            }
             $response = User::deleteItem($request, $id);
         } catch (\Exception $e) {
             $response = [];
